requirements.txt
astronomer-cosmos
apache-airflow-providers-snowflake
apache-airflow-providers-amazon
boto3==1.34.37
botocore==1.34.37
DateTime==5.4
pendulum==3.0.0
s3transfer
requests==2.31.0
pandas

